Restructure Alexa integration into smarthome/ and skill/ subfolders
Move all smarthome logic (request handlers, response helpers, discovery,
Alexa Events API client, types) into services/alexa/smarthome/, and all
skill logic (intent handlers, dispatch) into services/alexa/skill/. Both
routes now dispatch identically via an imported handler map. Dissolves
the requestTypes/ and intentTypes/ folder indirection.
