--- server/src/services/alexa/intentTypes/index.ts
@@ -1,2 +1,0 @@
-export { default as GoingOut } from './going-out';
-export { default as WhatsTheMessage } from './whats-the-message';
