+++ server/src/services/alexa/skill/types.ts
@@ -1,0 +1,9 @@
+export interface AlexaIntent {
+  name: string;
+  slots: Record<string, unknown>;
+}
+
+export interface AlexaSkillRequest {
+  type: string;
+  intent: AlexaIntent;
+}
