+++ server/src/services/alexa/smarthome/types.ts
